nnxn
This game is based on the game of life. It shows how humanity multiplies. 
The blue cells are male individuals, and the pink ones are female, respectively.

In the game, 4 cells of random color are randomly created, then the program reads the cells and stops when it finds blue cells, after which the program reads the cells located in the neighboring blue cells for the presence of pink cells. When the blue and pink cells touch, 4 more random-colored cells are generated in neighboring cells, and the old cells die. The game continues as long as the cells remain, let's say, without their pair, thereby ending the kind of life.